[FIX] #74 component separation
