+++ src/components/constants/chartStyles.js
@@ -1,0 +1,5 @@
+export const labelStyles = {
+  red: { color: 'red', fontWeight: 'bold', fontSize: 14 },
+  blue: { color: 'blue', fontWeight: 'bold', fontSize: 14 },
+  highlight: { color: '#000', fontWeight: 'bold', fontSize: 14 },
+}
